splits accounting db into storagenodeaccounting and projectaccounting interfaces and renames methods to match
