+++ pkg/accounting/projectusage_test.go
@@ -39,7 +39,7 @@ func TestProjectUsageStorage(t *testing.T) {
 		SatelliteCount: 1, StorageNodeCount: 6, UplinkCount: 1,
 	}, func(t *testing.T, ctx *testcontext.Context, planet *testplanet.Planet) {
 		saDB := planet.Satellites[0].DB
-		acctDB := saDB.Accounting()
+		acctDB := saDB.ProjectAccounting()
 
 		// Setup: create a new project to use the projectID
 		projects, err := planet.Satellites[0].DB.Console().Projects().GetAll(ctx)
@@ -57,7 +57,7 @@ func TestProjectUsageStorage(t *testing.T) {
 				}
 
 				// Execute test: get storage totals for a project, then check if that exceeds the max usage limit
-				inlineTotal, remoteTotal, err := acctDB.ProjectStorageTotals(ctx, projectID)
+				inlineTotal, remoteTotal, err := acctDB.GetStorageTotals(ctx, projectID)
 				require.NoError(t, err)
 				maxAlphaUsage := 25 * memory.GB
 				actualExceeded, actualResource := accounting.ExceedsAlphaUsage(0, inlineTotal, remoteTotal, maxAlphaUsage)
@@ -97,7 +97,7 @@ func TestProjectUsageBandwidth(t *testing.T) {
 	}, func(t *testing.T, ctx *testcontext.Context, planet *testplanet.Planet) {
 		saDB := planet.Satellites[0].DB
 		orderDB := saDB.Orders()
-		acctDB := saDB.Accounting()
+		acctDB := saDB.ProjectAccounting()
 
 		// Setup: get projectID and create bucketID
 		projects, err := planet.Satellites[0].DB.Console().Projects().GetAll(ctx)
@@ -128,7 +128,7 @@ func TestProjectUsageBandwidth(t *testing.T) {
 				from := time.Now().AddDate(0, 0, -accounting.AverageDaysInMonth)
 
 				// Execute test: get bandwidth totals for a project, then check if that exceeds the max usage limit
-				bandwidthTotal, err := acctDB.ProjectAllocatedBandwidthTotal(ctx, bucketID, from)
+				bandwidthTotal, err := acctDB.GetAllocatedBandwidthTotal(ctx, bucketID, from)
 				require.NoError(t, err)
 				maxAlphaUsage := 25 * memory.GB
 				actualExceeded, actualResource := accounting.ExceedsAlphaUsage(bandwidthTotal, 0, 0, maxAlphaUsage)
@@ -155,7 +155,7 @@ func createBucketID(projectID uuid.UUID, bucket []byte) []byte {
 	return []byte(storj.JoinPaths(entries...))
 }
 
-func setUpStorageTallies(ctx *testcontext.Context, projectID uuid.UUID, acctDB accounting.DB, time time.Time) error {
+func setUpStorageTallies(ctx *testcontext.Context, projectID uuid.UUID, acctDB accounting.ProjectAccounting, time time.Time) error {
 
 	// Create many records that sum greater than project usage limit of 25GB
 	for i := 0; i < 4; i++ {
@@ -169,7 +169,7 @@ func setUpStorageTallies(ctx *testcontext.Context, projectID uuid.UUID, acctDB a
 			// that sum greater than the maxAlphaUsage * expansionFactor
 			RemoteBytes: 10 * memory.GB.Int64() * accounting.ExpansionFactor,
 		}
-		err := acctDB.CreateBucketStorageTally(ctx, tally)
+		err := acctDB.CreateStorageTally(ctx, tally)
 		if err != nil {
 			return err
 		}
@@ -226,7 +226,7 @@ func TestProjectBandwidthTotal(t *testing.T) {
 		ctx := testcontext.New(t)
 		defer ctx.Cleanup()
 
-		accountingDB := db.Accounting()
+		pdb := db.ProjectAccounting()
 		projectID, err := uuid.New()
 		require.NoError(t, err)
 
@@ -237,7 +237,7 @@ func TestProjectBandwidthTotal(t *testing.T) {
 		// Execute test: get project bandwidth total
 		bucketID := createBucketID(*projectID, []byte("testbucket"))
 		from := time.Now().AddDate(0, 0, -accounting.AverageDaysInMonth) // past 30 days
-		actualBandwidthTotal, err := accountingDB.ProjectAllocatedBandwidthTotal(ctx, bucketID, from)
+		actualBandwidthTotal, err := pdb.GetAllocatedBandwidthTotal(ctx, bucketID, from)
 		require.NoError(t, err)
 		require.Equal(t, actualBandwidthTotal, expectedTotal)
 	})
