+++ pkg/accounting/rollup/rollup.go
@@ -26,16 +26,16 @@ type Config struct {
 type Service struct {
 	logger        *zap.Logger
 	ticker        *time.Ticker
-	db            accounting.DB
+	sdb           accounting.StoragenodeAccounting
 	deleteTallies bool
 }
 
 // New creates a new rollup service
-func New(logger *zap.Logger, db accounting.DB, interval time.Duration, deleteTallies bool) *Service {
+func New(logger *zap.Logger, sdb accounting.StoragenodeAccounting, interval time.Duration, deleteTallies bool) *Service {
 	return &Service{
 		logger:        logger,
 		ticker:        time.NewTicker(interval),
-		db:            db,
+		sdb:           sdb,
 		deleteTallies: deleteTallies,
 	}
 }
@@ -60,7 +60,7 @@ func (r *Service) Run(ctx context.Context) (err error) {
 // Rollup aggregates storage and bandwidth amounts for the time interval
 func (r *Service) Rollup(ctx context.Context) error {
 	// only Rollup new things - get LastRollup
-	lastRollup, err := r.db.LastTimestamp(ctx, accounting.LastRollup)
+	lastRollup, err := r.sdb.LastTimestamp(ctx, accounting.LastRollup)
 	if err != nil {
 		return Error.Wrap(err)
 	}
@@ -83,14 +83,14 @@ func (r *Service) Rollup(ctx context.Context) error {
 		return nil
 	}
 
-	err = r.db.SaveRollup(ctx, latestTally, rollupStats)
+	err = r.sdb.SaveRollup(ctx, latestTally, rollupStats)
 	if err != nil {
 		return Error.Wrap(err)
 	}
 
 	if r.deleteTallies {
 		// Delete already rolled up tallies
-		err = r.db.DeleteRawBefore(ctx, latestTally)
+		err = r.sdb.DeleteTalliesBefore(ctx, latestTally)
 		if err != nil {
 			return Error.Wrap(err)
 		}
@@ -101,7 +101,7 @@ func (r *Service) Rollup(ctx context.Context) error {
 
 // RollupStorage rolls up storage tally, modifies rollupStats map
 func (r *Service) RollupStorage(ctx context.Context, lastRollup time.Time, rollupStats accounting.RollupStats) (latestTally time.Time, err error) {
-	tallies, err := r.db.GetRawSince(ctx, lastRollup)
+	tallies, err := r.sdb.GetTalliesSince(ctx, lastRollup)
 	if err != nil {
 		return time.Now(), Error.Wrap(err)
 	}
@@ -126,12 +126,7 @@ func (r *Service) RollupStorage(ctx context.Context, lastRollup time.Time, rollu
 			rollupStats[iDay][node] = &accounting.Rollup{NodeID: node, StartTime: iDay}
 		}
 		//increment data at rest sum
-		switch tallyRow.DataType {
-		case accounting.AtRest:
-			rollupStats[iDay][node].AtRestTotal += tallyRow.DataTotal
-		default:
-			r.logger.Info("rollupStorage no longer supports non-accounting.AtRest datatypes")
-		}
+		rollupStats[iDay][node].AtRestTotal += tallyRow.DataTotal
 	}
 
 	return latestTally, nil
@@ -140,7 +135,7 @@ func (r *Service) RollupStorage(ctx context.Context, lastRollup time.Time, rollu
 // RollupBW aggregates the bandwidth rollups, modifies rollupStats map
 func (r *Service) RollupBW(ctx context.Context, lastRollup time.Time, rollupStats accounting.RollupStats) error {
 	var latestTally time.Time
-	bws, err := r.db.GetStoragenodeBandwidthSince(ctx, lastRollup.UTC())
+	bws, err := r.sdb.GetBandwidthSince(ctx, lastRollup.UTC())
 	if err != nil {
 		return Error.Wrap(err)
 	}
