+++ pkg/accounting/tally/tally.go
@@ -26,23 +26,25 @@ type Config struct {
 
 // Service is the tally service for data stored on each storage node
 type Service struct {
-	logger       *zap.Logger
-	metainfo     *metainfo.Service
-	overlay      *overlay.Cache
-	limit        int
-	ticker       *time.Ticker
-	accountingDB accounting.DB
+	logger                  *zap.Logger
+	metainfo                *metainfo.Service
+	overlay                 *overlay.Cache
+	limit                   int
+	ticker                  *time.Ticker
+	storagenodeAccountingDB accounting.StoragenodeAccounting
+	projectAccountingDB     accounting.ProjectAccounting
 }
 
 // New creates a new tally Service
-func New(logger *zap.Logger, accountingDB accounting.DB, metainfo *metainfo.Service, overlay *overlay.Cache, limit int, interval time.Duration) *Service {
+func New(logger *zap.Logger, sdb accounting.StoragenodeAccounting, pdb accounting.ProjectAccounting, metainfo *metainfo.Service, overlay *overlay.Cache, limit int, interval time.Duration) *Service {
 	return &Service{
-		logger:       logger,
-		metainfo:     metainfo,
-		overlay:      overlay,
-		limit:        limit,
-		ticker:       time.NewTicker(interval),
-		accountingDB: accountingDB,
+		logger:                  logger,
+		metainfo:                metainfo,
+		overlay:                 overlay,
+		limit:                   limit,
+		ticker:                  time.NewTicker(interval),
+		storagenodeAccountingDB: sdb,
+		projectAccountingDB:     pdb,
 	}
 }
 
@@ -71,13 +73,13 @@ func (t *Service) Tally(ctx context.Context) error {
 		errAtRest = errs.New("Query for data-at-rest failed : %v", err)
 	} else {
 		if len(nodeData) > 0 {
-			err = t.SaveAtRestRaw(ctx, latestTally, time.Now().UTC(), nodeData)
+			err = t.SaveTallies(ctx, latestTally, time.Now().UTC(), nodeData)
 			if err != nil {
 				errAtRest = errs.New("Saving storage node data-at-rest failed : %v", err)
 			}
 		}
 		if len(bucketData) > 0 {
-			_, err = t.accountingDB.SaveBucketTallies(ctx, latestTally, bucketData)
+			_, err = t.projectAccountingDB.SaveTallies(ctx, latestTally, bucketData)
 			if err != nil {
 				errBucketInfo = errs.New("Saving bucket storage data failed")
 			}
@@ -91,7 +93,7 @@ func (t *Service) Tally(ctx context.Context) error {
 func (t *Service) CalculateAtRestData(ctx context.Context) (latestTally time.Time, nodeData map[storj.NodeID]float64, bucketTallies map[string]*accounting.BucketTally, err error) {
 	defer mon.Task()(&ctx)(&err)
 
-	latestTally, err = t.accountingDB.LastTimestamp(ctx, accounting.LastAtRestTally)
+	latestTally, err = t.storagenodeAccountingDB.LastTimestamp(ctx, accounting.LastAtRestTally)
 	if err != nil {
 		return latestTally, nodeData, bucketTallies, Error.Wrap(err)
 	}
@@ -201,7 +203,7 @@ func (t *Service) CalculateAtRestData(ctx context.Context) (latestTally time.Tim
 	return latestTally, nodeData, bucketTallies, err
 }
 
-// SaveAtRestRaw records raw tallies of at-rest-data and updates the LastTimestamp
-func (t *Service) SaveAtRestRaw(ctx context.Context, latestTally time.Time, created time.Time, nodeData map[storj.NodeID]float64) error {
-	return t.accountingDB.SaveAtRestRaw(ctx, latestTally, created, nodeData)
+// SaveTallies records raw tallies of at-rest-data and updates the LastTimestamp
+func (t *Service) SaveTallies(ctx context.Context, latestTally time.Time, created time.Time, nodeData map[storj.NodeID]float64) error {
+	return t.storagenodeAccountingDB.SaveTallies(ctx, latestTally, created, nodeData)
 }
